[Chore] rename `fakeAsync.ts`
